+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased] - 2019-01-21
 ### Added
  - Add a `-c` parameter to run shell commands in the game's installation directory with relevant Wine environment variables.
+ - Non-Steam applications are now detected.
 
 ### Fixed
  - All arguments are now correctly passed to winetricks.